+++ src/trtutils/__init__.py
@@ -15,11 +15,19 @@ backends
     A module providing alternative backends for the TRTEngine class.
 core
     A module for the core functionality of the package.
+jetson
+    A module implementating additional functionality for Jetson devices.
+impls
+    A module containing implementations for different neural networks.
 trtexec
     A module for utilities related to the trtexec tool.
 
 Classes
 -------
+BenchmarkResult
+    A dataclass for storing profiling information from benchmarking engines.
+Metric
+    A dataclass storing specific metric information from benchmarking.
 TRTEngine
     A class for creating TensorRT engines from serialized engine files.
 TRTModel
@@ -35,8 +43,15 @@ QueuedTRTModel
 
 Functions
 ---------
+benchmark_engine
+    Benchmark a TensorRT engine.
 find_trtexec
     Find an instance of the trtexec binary on the system.
+run_trtexec
+    Run a command with trtexec.
+set_log_level
+    Set the log level of the trtutils package.
+
 """
 
 from __future__ import annotations
@@ -110,15 +125,17 @@ if level is not None and level.upper() not in [
     _log.warning(f"Invalid log level: {level}. Using default log level: WARNING")
 
 __author__ = "Justin Davis"
-__version__ = "0.1.2"
+__version__ = "0.2.0"
 
-
-from . import backends, core, trtexec
+from . import backends, core, impls, jetson, trtexec
+from ._benchmark import BenchmarkResult, Metric, benchmark_engine
 from ._engine import ParallelTRTEngines, QueuedTRTEngine, TRTEngine
 from ._model import ParallelTRTModels, QueuedTRTModel, TRTModel
 from .trtexec import find_trtexec, run_trtexec
 
 __all__ = [
+    "BenchmarkResult",
+    "Metric",
     "ParallelTRTEngines",
     "ParallelTRTModels",
     "QueuedTRTEngine",
@@ -126,8 +143,11 @@ __all__ = [
     "TRTEngine",
     "TRTModel",
     "backends",
+    "benchmark_engine",
     "core",
     "find_trtexec",
+    "impls",
+    "jetson",
     "run_trtexec",
     "set_log_level",
     "trtexec",