+++ src/trtutils/_model.py
@@ -73,14 +73,43 @@ class TRTModel:
             warmup_iterations=warmup_iterations,
             warmup=warmup,
         )
-        self._preprocess = preprocess
-        self._postprocess = postprocess
+        self._preprocess: Callable[[list[np.ndarray]], list[np.ndarray]] = preprocess
+        self._postprocess: Callable[[list[np.ndarray]], list[np.ndarray]] = postprocess
+
+    @property
+    def engine(self: Self) -> TRTEngine:
+        """Access the underlying TRTEngine."""
+        return self._engine
+
+    @property
+    def preprocessor(self: Self) -> Callable[[list[np.ndarray]], list[np.ndarray]]:
+        """The preprocessing function used in this model."""
+        return self._preprocess
+
+    @preprocessor.setter
+    def preprocessor(
+        self: Self,
+        new_preprocess: Callable[[list[np.ndarray]], list[np.ndarray]],
+    ) -> None:
+        """
+        Set the preprocessing function used in this model.
+
+        Useful in case the preprocessor need information which is only
+        accessible after loading the engine.
+        """
+        self._preprocess = new_preprocess
+
+    @property
+    def postprocessor(self: Self) -> Callable[[list[np.ndarray]], list[np.ndarray]]:
+        """The postprocessing function used in this model."""
+        return self._postprocess
 
     def __call__(
         self: Self,
         inputs: list[np.ndarray],
         *,
         preprocessed: bool | None = None,
+        postprocess: bool | None = None,
     ) -> list[np.ndarray]:
         """
         Execute the model with the given inputs.
@@ -92,7 +121,9 @@ class TRTModel:
         preprocessed : bool, optional
             Whether the inputs are already preprocessed, by default None
             If None, the inputs will be preprocessed
-
+        postprocess : bool, optional
+            Whether or not to postprocess the outputs, by default None
+            If None, the outputs will be postprocessed
 
         Returns
         -------
@@ -100,13 +131,11 @@ class TRTModel:
             The outputs of the model
 
         """
-        return self.run(inputs, preprocessed=preprocessed)
+        return self.run(inputs, preprocessed=preprocessed, postprocess=postprocess)
 
     def mock_run(
         self: Self,
         data: list[np.ndarray] | None = None,
-        *,
-        preprocessed: bool | None = None,
     ) -> list[np.ndarray]:
         """
         Execute the model with random inputs.
@@ -116,9 +145,6 @@ class TRTModel:
         data : list[np.ndarray], optional
             The inputs to the model, by default None
             If None, random inputs will be used
-        preprocessed : bool, optional
-            Whether the inputs are already preprocessed, by default None
-            If None, the inputs will be preprocessed.
 
         Returns
         -------
@@ -128,8 +154,7 @@ class TRTModel:
         """
         if data is None:
             data = self._engine.get_random_input()
-        outputs = self.run(data, preprocessed=preprocessed)
-        return self._postprocess(outputs)
+        return self.run(data, preprocessed=True, postprocess=False)
 
     def preprocess(self: Self, inputs: list[np.ndarray]) -> list[np.ndarray]:
         """
@@ -148,11 +173,29 @@ class TRTModel:
         """
         return self._preprocess(inputs)
 
+    def postprocess(self: Self, outputs: list[np.ndarray]) -> list[np.ndarray]:
+        """
+        Postprocess the outputs.
+
+        Parameters
+        ----------
+        outputs : list[np.ndarray]
+            The outputs to postprocess
+
+        Returns
+        -------
+        list[np.ndarray]
+            The postprocessed outputs
+
+        """
+        return self._postprocess(outputs)
+
     def run(
         self: Self,
         inputs: list[np.ndarray],
         *,
         preprocessed: bool | None = None,
+        postprocess: bool | None = None,
     ) -> list[np.ndarray]:
         """
         Execute the model with the given inputs.
@@ -164,7 +207,9 @@ class TRTModel:
         preprocessed : bool, optional
             Whether the inputs are already preprocessed, by default None
             If None, the inputs will be preprocessed
-
+        postprocess : bool, optional
+            Whether or not to postprocess the outputs, by default None
+            If None, the outputs will be postprocessed
 
         Returns
         -------
@@ -174,10 +219,14 @@ class TRTModel:
         """
         if preprocessed is None:
             preprocessed = False
+        if postprocess is None:
+            postprocess = True
         if not preprocessed:
             inputs = self._preprocess(inputs)
         outputs = self._engine.execute(inputs)
-        return self._postprocess(outputs)
+        if postprocess:
+            outputs = self._postprocess(outputs)
+        return outputs
 
 
 class QueuedTRTModel: