+++ src/trtutils/jetson/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2024 Justin Davis (davisjustin302@gmail.com)
+#
+# MIT License
+"""A submodule implementing additional tools for Jetson devices."""
+
+from __future__ import annotations